requirements.txt
streamlit
pandas
requests
beautifulsoup4
tqdm
biopython
feedparser